export individual configs in index
